Loa Django/requirements.txt
django-tables2
django-2.2.2
django-bootstrap4
